refactor: add useHeader for reusing logics for detail and header
